remove pre-save of objects before autonumber.
This was needed to allow business rules to run that use other tables
before the target table was locked because in mysql you can only use
tables mentioned in the lock statement. But this causes problems when
the autonumber field can't be null. Right now, there don't seem to be
any business rules that causing problems, so I'm removing this. It might
be posible to override the save method somehow so that it locks the
table after the businessrules are done... Maybe there is a better way to
autonumber. Fixes #195