store the proper MIME type for PIA assets
used to store the MIME type of the uploaded file, but the MIME type would differ if upload requires file conversion (e.g. if image server used is iipsrv which requires all files to be in tiff).

Will now store the MIME type returned by PIA for the file as stored on the server