adjust mysql transaction isolation level during wb row validation.
In the default isolation level of READ COMMITED deadlocks can occur
when multiple row validation request occur concurrently. It seems to
be related uniqueness constraints on a field. For instance, using
fill-down in a dataset to set catalog number to the same value for
several rows results in some requests failing due to mysql
deadlock. See the comment about duplicate-key errors in the INSERT
section at
https://dev.mysql.com/doc/refman/8.0/en/innodb-locks-set.html.

Setting the isolation level to READ UNCOMMITTED allows the
validation requests to succeed. Since this change only affects live
validation which always rolls back any changes, the possibility of
introducing inconsistencies can be ignored.