Make front-end's caching module more type safe
The names of the buckets, the cache names and the types of each of them
are now defined in a separate file.

Trying to read or write to an undefined cache results in a TypeScript
type error.

Still, if for some reason a generic variant is needed, there exists
`genericGet` and `genericSet`.