Fix wbplanview converting rank names to upper case
WbPlanView made an assumption that all rank names are in upper case,
yet, for simplicity it was internally converting all tree ranks to lower
case (because that's what it does with field names).
Also, it was ignoring the "Title" attribute on tree ranks and was
instead showing the "name" attribute.
This commit fixes those issues. AutoMapper and AutoMapperDefinitions
still refer to ranks by their lower case name, but they make sure to
convert them back before returning the results. The other parts of
WbPlanView are now handling tree ranks as they are, without converting
them to upper case or lower case.