Remove `isModified: persistent` cell meta state
The state was used to identify disambiguated cells as those should be
marked as isModified even though cell's value did not change.
Though, this caused issue of isModified getting out of sync on
undoing/redoing disambiguation.
Checking whether the cell was disambiguated from inside of setCellMeta
allows us to get rid of the the `persistent` state.