Partial workaround for onPaste into input[type="date"] in FF
input[type="date"] and input[type="month"] fields do not support pasting
values into them by default.

Adding onPaste event listen in Google Chrome seems to be a good
workaround, though, that does not work in Firefox as it does not seem
to emit an onPaste event for these input fields.

Adding event listener to the parent element seems to make it work in
Firefox but only some of the time, for some weird reason.

The only consistent workaround for pasting in Firefox is to click on
the background between the <select> and <input type="date"> fields and
then to press Ctrl+V. Sigh...

Another solution would be to replace the input's type with text on key
down and return back on key up in anticipation of Ctrl+V, though that
sounds like a very clunky solution.