Use aria-label correctly
aria-label is used to replace the text context of an element in the
accessibility tree. It is to be used when an element does not have a
<label> and does not have text (or descriptive) text inside.

For cases where the label is self-explanatory, but additional
information needs to be given, use [aria-description] attribute. It is
not yet supported by all assistive tools, thus [title] attribute has been
used too, where appropriate.

Additionally, site's logo is now wrapped in an <h1> and pages start
headers at <h2>