Fix potential bugs with localization string usages
If localization strings accepted arguments, the call to the
dictionary returned a function that in turn accepted arguments.
However, it was possible to forget to call the returned function
with arguments and instead use the function itself.
This was especially a problem in the context of React components, which
do not throw a compile time type error about trying to render a
function instead of JSX.Element/string. However, that thew an error
on the runtime.

This commit changes "createDictionary" function to expect all arguments
to a localization string, if any in the first function call.

Before the change:

```js
commonText('newResourceTitle')(label)
```

After the change:

```js
commonText('newResourceTitle', label)
```

This error was catchable by localization tests
(./tests/testlocalization.ts), but the new method is better because
an error is caught even earlier.