Make query's "Form View" dialog be full height
Making the dialog always occupy full height fixed the problem of the
next/previous record buttons chaning their position on the screen as
a result of the forms for different records varying in height, and
thus causing dialog size to change.

As an additional side effect, this change fixes the form footer buttons
to the bottom of the screen.