Replace usages of "window" with "globalThis"
globalThis is equivalent to window in modern browsers.

It also provides additional benefit of working in Node.js enviroments,
(which is the case when running tests)

Since `window` object is not defined in tests, any code that relied
on window had to check if window is defined prior to using it or
risk blowing up when executed as part of the test. Now, that is no
longer necessary