add api to create RS from WB upload instead of just doing it.
This commit removes the automatic creation of a record set of uploaded
records during a data set upload.

It adds a new endpoint /api/workbench/create_recordset/{ds_id}/ that
accepts a name parameter and creates a record set of that name from
the upload results of data set with the given id.

It adds a new permission check for resource = '/workbench/dataset'
action = 'create_recordset'.

It adds a data migration to automatically grant said permission to any
user or role currently possessing permission on resource
'/workbench/dataset' action 'upload'.