Add backward compatibility with pick lists in Sp6 QB
Sp7 QB stores PickListItem.value when selecting a pick list item as a
column filter

Sp6 QB uses PickListItem.title column, which is probably a bug or
an oversight as that column is intended for UI purposes only.

This commit makes Sp7 UI be able to resolve PickListItem even if it
was saved using the title field. The value would still be saved using
the value field, as that is the correct behavior.

Fixes #1934