Rename "Model" to "Table"
Renames all occurrences of "Model" with "Table"

"Model" was renamed to "Table" because:
- "Model" was used to refer both to table and a resource of that table
- "Model" is related to Django Models and makes sense on the back-end
  but makes little sense on the front-end. Backbone uses term "Model",
  but we are planning on getting rid of Backbone.
- Front-end has to print errors to users. Errors should reference
  "Table" not "Model" as users have no idea what "Model" means.
- Both "Model" and "Table" were often used interchangeably. Having just
  one term adds consistency.

Still to do:
- Rename the remaining usages of "model"
- Get rid of "record" or "resource" in favor of the other for
  consistency.

Also, untangles the "schema.ts" and "schemaBase.ts" files. They make
separate network requests so it doesn't make sense to store their
results in the same object. Additionally, commonly just one of them is
needed.

Instead, `schema.models` is renamed to `tables`. The rest of the
properties on `schema` remain the same, except `schemaBase` is renamed
to `schema`.