Make Syncer more resilient to unknown attributes
There was an edge case that wasn't handled properly:
Let's say there are 2 children nodes of some node.
1st child includes unknown attributes (attributes that are not in a
spec) or unknown children
Later, first child is deleted in JSON and JSON is converted back to
XML.
As a result, the unknown attributes and unknown children would be
carried over from 1st child to 2nd.
Any unknown attributes on the 2nd child would get lost - not good.

Because of this problem, any attribute not defined in the spec (i.e,
because Sp7 does not use it, or because it was mistyped) would be
mismatched when node index changed.

This gets even worse in the FormEditor spec as that includes
syncer.switch, where unknown attributes for one node might not be
unknown attributes for another node

Fixed all of that by keeping track of the original attributes on a node
in a symbol, and having that symbol be attached to every node.