Don't trigger unload protect when setting lat/long
Fixes #3584

The resource.set() function was setting an unload protect when
intially loading lats and longs that were very precise.

The intial value (string), was being set to a number which was
not equal to the inital value.

This led records to need saving even when they were just loaded
(from QueryBuilder or other means)