Split dataset views file to prevent readding of dataset permission
Looks like Django tries to continue loading the rest of the file in case of an error.
So, if fetching asset server fails the first time specify starts,
the permission of attachment dataset will be added again (since it was located in the same file as the file with error),
which lead to duplicate addition of permission in permission registry, which raises an error.
Splitting the file prevents to continuation in case of error.