Correctyl send notifications for sending RSS Export Feeds
Implements the backend changes for #4267

If the notifyUserId is specified in the ExportFeed, always send the
full notification to that user. Send a notification to the
requesting user (without file) when the process completes.

If notifyUserId is not specified, then send the full export feed
notification to the requesting user.