fix(map): separate spnetwork and query contexts
Both queries and the specify network feature utilize the same component
and the same descriptions for showing records on a map. This fix checks
to see if there is a gbif or idigbio description available (as there
would be in the specify network context) and only shows the details
dropdown in that case. If the user is viewing the map from query
results, no description is shown, as it is not accurate in that context.

Additionally, adds a space between the descriptions if more than one.