feat: move stats collection out of initialContext
Previously, stats would block the application until it was
resolved. This led to a serious issue in which the stats
server was not responding and the request timed out: the
application would be completely inacessible while this was
occuring.
Because it's not essential, stats are now sent in
the background.